+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog Format
 - [Initials of Dev] [#issue number](issue link on GitHub) {issue title}
 -->
 
+- [MD] [#69](https://github.com/Ateneo-MISA/MISA-Website/issues/69) vote(feature): add elections and candidates pages
 - [KF] [#64](https://github.com/Ateneo-MISA/MISA-Website/issues/64) app(feature): add environment to config
 - [KF] [#61](https://github.com/Ateneo-MISA/MISA-Website/issues/61) app(feature): add and reorganize contentful queries
 - [KF] [#58](https://github.com/Ateneo-MISA/MISA-Website/issues/58) home page(feature): fix upcoming events layout
@@ -33,5 +34,6 @@ Changelog Format
 - [KF] [#9](https://github.com/Ateneo-MISA/MISA-Website/issues/9) app(feature): add tailwind css
 - [KF MD] [#5](https://github.com/Ateneo-MISA/MISA-Website/issues/5) home page(feature): add events frame
 - [MF] [#4](https://github.com/Ateneo-MISA/MISA-Website/issues/4) home page(feature): add hero
+- [AL] [#3](https://github.com/Ateneo-MISA/MISA-Website/issues/3) home page(feature): add footer
 - [KF] [#2](https://github.com/Ateneo-MISA/MISA-Website/issues/2) home page(feature): add navbar
 - [KF] [#1](https://github.com/Ateneo-MISA/MISA-Website/issues/1) app(feature): add initial project files